+++ models/models.go
@@ -6,10 +6,10 @@ import (
 )
 
 type CoreUser struct {
-	ID        uint                    `gorm:"primarykey" json:"-"`
-	UserID    uint                    `gorm:"uniqueIndex:index_unique_core_user" json:"user_id"`
+	ID        uint64                  `gorm:"primarykey" json:"-"`
+	UserID    uint64                  `gorm:"uniqueIndex:index_unique_core_user" json:"user_id"`
 	User      utilitymodels.LocalUser `json:"-"`
-	CoreID    uint                    `gorm:"uniqueIndex:index_unique_core_user" json:"core_id"`
+	CoreID    uint64                  `gorm:"uniqueIndex:index_unique_core_user" json:"core_id"`
 	CreatedAt time.Time               `json:"-"`
 	UpdatedAt time.Time               `json:"-"`
 }
